requirements.txt
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gspread==6.2.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl
pandas